+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Add featured packs tab
 - Add message to accounts tab with info on what account to use
 - Fix loading Forge loader on older MC versions
+- Update default ram and java arguments and remove warning when allocating more than 4GB of ram